+++ updater/software/All.cs
@@ -1,6 +1,6 @@
 ﻿/*
     This file is part of the updater command line interface.
-    Copyright (C) 2017, 2018, 2019, 2021, 2022, 2023, 2024  Dirk Stolle
+    Copyright (C) 2017, 2018, 2019, 2021, 2022, 2023, 2024, 2025  Dirk Stolle
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -101,6 +101,8 @@ namespace updater.software
             result.Add(new MariaDB_10_6(autoGetNewer));
             result.Add(new MariaDB_10_11(autoGetNewer));
             result.Add(new MariaDB_11_4(autoGetNewer));
+            result.Add(new VisualCppRedist_2015_2022_x86(autoGetNewer));
+            result.Add(new VisualCppRedist_2015_2022_x64(autoGetNewer));
             result.Add(new Mumble(autoGetNewer));
             result.Add(new NodeJS(autoGetNewer));
             result.Add(new NotepadPlusPlus(autoGetNewer));